+++ apps/demo/src/app/app.component.ts
@@ -1,7 +1,9 @@
-import { Component, ElementRef, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { emojiExampleList } from './example-emoji-list';
-import { BehaviorSubject, Subject, Subscription } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
+
+const exampleMD = () => fetch('./assets/example.md').then(r => r.text());
 
 @Component({
   selector: 'gewd-utils-root',
@@ -12,6 +14,7 @@ export class AppComponent implements OnInit, OnDestroy {
   private markdown$ = new Subject();
   private subscription: Subscription;
 
+  exampleMD = exampleMD();
   emojiList = emojiExampleList;
 
   @ViewChild('markdown', {static: true})