+++ apps/demo/src/app/app.module.ts
@@ -9,7 +9,12 @@ import { FormsModule } from '@angular/forms';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatTabsModule } from '@angular/material/tabs';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { GetWorkerPayload, LoadMarkdownWorkerInjectorToken, LoadMermaidInjectorToken } from '@gewd/markdown/contracts';
+import {
+  DEFAULT_PRISM_OPTIONS,
+  GetWorkerPayload,
+  LoadMarkdownWorkerInjectorToken,
+  LoadMermaidInjectorToken
+} from '@gewd/markdown/contracts';
 
 const marked = () => new Worker('./markdown.worker.ts', { type: 'module' });
 
@@ -25,7 +30,19 @@ const marked = () => new Worker('./markdown.worker.ts', { type: 'module' });
     {
       provide: LoadMarkdownWorkerInjectorToken,
       useValue: {
-        getWorker: marked
+        getWorker: marked,
+        options: {
+          prism: {
+            ...DEFAULT_PRISM_OPTIONS,
+
+            /** if needed **/
+            languageFileType: 'min.js',  // if you want to use the minified assets
+            languageMap: {               // alias to load the real file
+              ts: 'typescript',          // default
+              cs: 'csharp'               // additional
+            }
+          }
+        }
       } as GetWorkerPayload
     },
     {