+++ libs/markdown/contracts/worker.interface.ts
@@ -1,3 +1,5 @@
+import { InjectionToken } from '@angular/core';
+
 export interface PrismOptions {
   languageMap: {[key: string]: string};
   languageFileType: string;
@@ -13,3 +15,19 @@ export interface MarkdownWorker {
   init(options: WorkerOptions);
   compile(str: string): string|Promise<string>;
 }
+
+export interface GetWorkerPayload {
+  getWorker: () => Worker;
+  options: WorkerOptions;
+}
+
+export const DEFAULT_PRISM_OPTIONS: PrismOptions = {
+  assetPath: '/assets/prism/',
+  languageFileType: 'js',
+  languageMap: {
+    ts: 'typescript'
+  }
+};
+
+export const LoadMarkdownWorkerInjectorToken = new InjectionToken<GetWorkerPayload>('LoadMarkdownWorkerInjectorToken');
+export const LoadMermaidInjectorToken = new InjectionToken<string>('LoadMermaidInjectorToken');