+++ libs/markdown/service/public_api.ts
@@ -1,1 +1,1 @@
-export * from './worker.service';
+export * from './markdown.service';